change PPTransaction amount to int8

+++ transactions.go
@@ -156,7 +156,7 @@ func (t ItemTransaction) Mutate(b *Battle) {
 // A transaction to change the PP of a move.
 type PPTransaction struct {
 	Move   *Move
-	Amount int
+	Amount int8
 }
 
 func (t PPTransaction) Mutate(b *Battle) {
